Migrate experiments test
